Added migration BD

+++ src/main/java/com/vbaslak/issuetracker/domain/Comment.java
@@ -1,16 +1,20 @@
 package com.vbaslak.issuetracker.domain;
 
 import net.bytebuddy.implementation.bind.annotation.Empty;
+import org.hibernate.validator.constraints.Length;
 
 import javax.persistence.*;
+import javax.validation.constraints.NotBlank;
 import java.util.Date;
 
 @Entity
 public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-
+    private Long id;
+//    Later
+//    @NotBlank(message = "Please fill out a text comment")
+//    @Length(max = 2048, message = "description too long (more than 2kB)")
     private String textComment;
     private String status;
     private Date dateComment;
@@ -34,11 +38,11 @@ public class Comment {
         this.dateComment = new Date();
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
